Refactor: Change location of constants
